organize pages and components files
